fix all style problem
