+++ fastchat/serve/inference.py
@@ -10,18 +10,10 @@ except ImportError:
     from transformers import AutoTokenizer, AutoModelForCausalLM, LLaMATokenizer, LLamaForCausalLM, AutoModel
 
 from fastchat.conversation import conv_templates, get_default_conv_template, SeparatorStyle
-from fastchat.serve.compression import compress_module, compress, CompressionConfig, get_compressed_list, apply_compressed_weight
+from fastchat.serve.compression import  load_compress_model
 from fastchat.serve.monkey_patch_non_inplace import replace_llama_attn_with_non_inplace_operations
 from fastchat.serve.serve_chatglm import chatglm_generate_stream
-import glob
-import os
-import gc
-from tqdm import tqdm
-from accelerate import init_empty_weights, load_checkpoint_and_dispatch
-from accelerate.utils import set_module_tensor_to_device
 
-default_compression_config = CompressionConfig(
-    num_bits=8, group_size=256, group_dim=1, symmetric=True, enabled=True)
 
 
 def raise_warning_for_old_weights(model_path, model):
@@ -104,44 +96,6 @@ def load_model(model_path, device, num_gpus, max_gpu_memory="13GiB",
 
     return model, tokenizer
 
-
-def load_compress_model(model_path, device):
-
-    # partially load model
-    tokenizer = AutoTokenizer.from_pretrained(model_path, use_fast=False)
-    base_pattern = os.path.join(model_path, "pytorch_model-*.bin")
-    files = glob.glob(base_pattern)
-    
-    with init_empty_weights():
-        config = AutoConfig.from_pretrained(model_path, low_cpu_mem_usage=True, device_map='auto',torch_dtype=torch.float16)
-        model = AutoModelForCausalLM.from_config(config)
-        linear_weights = get_compressed_list(model)
-
-
-    compressed_state_dict = {}
-
-    for file in files:
-        tmp_state_dict = torch.load(file, map_location="cpu")
-        for name in tqdm(tmp_state_dict):
-            if name in linear_weights:
-                compressed_state_dict[name] = tmp_state_dict[name].to(device)
-                tensor = compressed_state_dict[name].data
-                compressed_state_dict[name] = compress(tensor, default_compression_config)
-            else:
-                compressed_state_dict[name] = tmp_state_dict[name].to(device)
-            tmp_state_dict[name] = None
-            tensor = None
-            gc.collect()
-            torch.cuda.empty_cache()
-
-    for name in model.state_dict():
-        if name not in linear_weights:
-            set_module_tensor_to_device(model, name, device, value=compressed_state_dict[name])
-    apply_compressed_weight(model, compressed_state_dict, device)
-
-    model.to(device)
-
-    return model, tokenizer
 
 @torch.inference_mode()
 def generate_stream(model, tokenizer, params, device,
